requirements.txt
# API dependencies
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
python-dotenv==1.0.0

# Celery and Redis
celery==5.2.7
redis==4.5.5

# AI Model dependencies
onnxruntime==1.20.0
opencv-python==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
torch==2.0.1

# S3 and AWS
boto3==1.26.135

# Utilities
requests==2.30.0
